Update repository name in project
It's clearer to use a keyword to denote the project is an implementation for Jekyll, and a more appropriate use of third-party names. To avoid confusion with the formal `for` keyword, it's crucial to keep the project versions as development releases (`0.x.y`), since there is no formal spec to implement.