Drop cmd bundle and set Jekyll env vars
The program doesn't find the `bundle` or `jekyll` binaries when running from Kotlin, so it needs their paths to call them. Further, it doesn't find `gems` when running `bundle` even after passing the environment paths, so removing it and calling `jekyll` directly simplifies and fixes the execution. Thrid, Jekyll will fail to build out of a Git repository because of a GitHub Pages plugin, that's why the `PAGES_REPO_NWO` variable is set with a dummy value.